notification component frontend made and added to allcards and menu sidebar
